Backend/requirements.txt
fastapi==0.116.1
Pillow==11.3.0
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.5
sendgrid==6.12.4
SQLAlchemy==2.0.31
stripe==12.4.0
